chatbot-api/requirements.txt
fastapi
pydantic
langchain_openai
langchain_core
langchain
langchain_community
python-dotenv
apache-airflow==2.7.1
pymysql
PyMySQL==1.1.0
